Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
email-validator
fastapi-mail
